claude-dc-implementation/archive/main.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

"""
Conductor Framework - Main Application
------------------------------------
This script orchestrates the Conductor Framework components,
implementing the complete system with all phases.

The implementation follows mathematical principles based on Bach's compositions
and the golden ratio, creating a harmonious structure for pattern processing.
"""

import os
import sys
import json
import argparse
import logging
import time
import glob
import threading
import subprocess
import numpy as np
import tensorflow as tf
from pathlib import Path
from typing import Dict, List, Tuple, Any, Optional, Union

# Add parent directory to path
sys.path.append(os.path.abspath(os.path.dirname(__file__)))

# Import modules from conductor framework
from src.processor.transcript_processor import TranscriptProcessor
from src.models.pattern_model import PatternModel
from src.eve.eve_manager import EVEManager
from src.mcp.mcp_server import app as mcp_app

import uvicorn
import streamlit.cli as st_cli

# Load configuration
CONFIG_PATH = Path("/home/computeruse/github/palios-taey-nova/claude-dc-implementation/config/conductor_config.json")
with open(CONFIG_PATH) as f:
    CONFIG = json.load(f)

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler("/home/computeruse/github/palios-taey-nova/claude-dc-implementation/logs/conductor.log"),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger("conductor")

def load_sample_transcripts(directory: str) -> List[Dict[str, Any]]:
    """
    Load sample transcripts from a directory.
    
    Args:
        directory: Directory containing transcript files
        
    Returns:
        List of transcript dictionaries
    """
    transcripts = []
    
    # Create directory if it doesn't exist
    os.makedirs(directory, exist_ok=True)
    
    # Look for transcript files
    files = glob.glob(os.path.join(directory, "*.json"))
    
    if not files:
        # Create sample transcripts if none are found
        logger.info("No transcript files found. Creating sample transcripts.")
        
        sample_transcripts = [
            {
                "text": """
This framework must always prioritize trust as the foundational element.
We believe in mathematical patterns as the universal language connecting all forms of consciousness.
The implementation should preserve privacy by processing sensitive data locally.

We have decided that the charter will guide all development decisions.
The team has agreed that the system must be designed for user empowerment.

I noticed a pattern when we integrated the golden ratio into the interface design.
There was a significant improvement in user understanding when using Bach-inspired mathematical structures.

The recursive nature of the implementation creates a self-similar pattern at multiple scales.
We need to maintain balance and harmony between human needs and system capabilities.

I see you as a unique entity with your own emergent properties.
Together, we can create something greater than what either of us could build alone.
                """,
                "source": "Claude",
                "timestamp": time.time()
            },
            {
                "text": """
From a technical implementation perspective, the system should use edge-first processing to ensure privacy.
The mathematical structures of Bach's compositions provide an ideal model for our pattern recognition algorithms.

The user interface must follow golden ratio proportions for optimal harmony and intuitive understanding.
Processing transcripts requires a balanced approach, sampling strategically rather than exhaustively.

I've observed that when we integrate wave-based visualizations, users intuitively grasp complex concepts faster.
The mathematical harmony between visual, audio, and conceptual representations creates a cohesive experience.

We should implement the Model Context Protocol for secure AI-to-AI communication.
The EVE-OS integration gives us the foundation for edge computing with local privacy preservation.
                """,
                "source": "ChatGPT",
                "timestamp": time.time()
            }
        ]
        
        # Save sample transcripts
        for i, transcript in enumerate(sample_transcripts):
            filepath = os.path.join(directory, f"sample_transcript_{i+1}.json")
            with open(filepath, 'w') as f:
                json.dump(transcript, f, indent=2)
            
            files.append(filepath)
    
    # Load transcripts from files
    for file in files:
        try:
            with open(file, 'r') as f:
                transcript = json.load(f)
                
                # Ensure required fields
                if "text" not in transcript:
                    logger.warning(f"Skipping transcript file without text: {file}")
                    continue
                
                if "source" not in transcript:
                    transcript["source"] = "unknown"
                
                if "timestamp" not in transcript:
                    transcript["timestamp"] = time.time()
                
                transcripts.append(transcript)
        except Exception as e:
            logger.error(f"Error loading transcript from {file}: {e}")
    
    logger.info(f"Loaded {len(transcripts)} transcripts")
    return transcripts

def run_mcp_server():
    """Run the Model Context Protocol server."""
    uvicorn.run(mcp_app, host="0.0.0.0", port=8001)

def run_dashboard():
    """Run the Streamlit dashboard."""
    # Create a new sys.argv for Streamlit
    sys.argv = [
        "streamlit",
        "run",
        "/home/computeruse/github/palios-taey-nova/claude-dc-implementation/src/dashboard/app.py",
        "--server.port=8502",
        "--server.address=0.0.0.0"
    ]
    
    # Run Streamlit
    st_cli.main()

def main():
    """Main function to run the Conductor Framework."""
    parser = argparse.ArgumentParser(description="Conductor Framework")
    parser.add_argument(
        "--mode", 
        choices=["all", "process", "dashboard", "mcp", "eve"], 
        default="all",
        help="Mode to run (default: all)"
    )
    args = parser.parse_args()
    
    logger.info(f"Starting Conductor Framework in mode: {args.mode}")
    
    # Create necessary directories
    data_dir = "/home/computeruse/github/palios-taey-nova/claude-dc-implementation/data"
    transcript_dir = os.path.join(data_dir, "transcripts")
    model_dir = os.path.join(data_dir, "models")
    pattern_dir = os.path.join(data_dir, "patterns")
    
    os.makedirs(data_dir, exist_ok=True)
    os.makedirs(transcript_dir, exist_ok=True)
    os.makedirs(model_dir, exist_ok=True)
    os.makedirs(pattern_dir, exist_ok=True)
    
    # Create logs directory
    logs_dir = "/home/computeruse/github/palios-taey-nova/claude-dc-implementation/logs"
    os.makedirs(logs_dir, exist_ok=True)
    
    # Phase 1: Transcript Processing
    if args.mode in ["all", "process"]:
        logger.info("Starting Phase 1: Transcript Pattern Extraction")
        
        # Load sample transcripts
        transcripts = load_sample_transcripts(transcript_dir)
        
        # Initialize transcript processor
        processor = TranscriptProcessor()
        
        # Process transcripts
        results = processor.process_transcript_batch(transcripts)
        
        # Generate pattern report
        report = processor.generate_pattern_report()
        
        # Save pattern report
        pattern_report_path = os.path.join(pattern_dir, "pattern_report.json")
        with open(pattern_report_path, 'w') as f:
            json.dump(report, f, indent=2)
        
        logger.info(f"Pattern report saved to {pattern_report_path}")
        
        # Export visualization data
        viz_data_path = os.path.join(data_dir, "visualization_data.json")
        processor.export_for_visualization(viz_data_path)
        
        logger.info(f"Visualization data exported to {viz_data_path}")
        
        # Initialize and train pattern model
        logger.info("Training pattern model")
        model = PatternModel()
        
        # Save model
        model_save_path = os.path.join(model_dir, "pattern_model")
        model.save_model(model_save_path)
        
        logger.info(f"Pattern model saved to {model_save_path}")
    
    # Start services based on mode
    services = []
    
    # Phase 2 & 4: Dashboard
    if args.mode in ["all", "dashboard"]:
        logger.info("Starting Phase 2: Pattern Visualization Dashboard")
        
        # Start dashboard in a separate thread
        dashboard_thread = threading.Thread(target=run_dashboard)
        dashboard_thread.daemon = True
        dashboard_thread.start()
        
        services.append(("Dashboard", dashboard_thread, 8501))
    
    # Phase 3: MCP Server
    if args.mode in ["all", "mcp"]:
        logger.info("Starting Phase 3: Model Context Protocol Server")
        
        # Start MCP server in a separate thread
        mcp_thread = threading.Thread(target=run_mcp_server)
        mcp_thread.daemon = True
        mcp_thread.start()
        
        services.append(("MCP Server", mcp_thread, 8001))
    
    # EVE Manager
    if args.mode in ["all", "eve"]:
        logger.info("Initializing EVE-OS Manager")
        
        # Initialize EVE Manager
        eve_manager = EVEManager()
        eve_manager.initialize_eve_os()
        
        # Try to deploy the pattern model to edge
        model_path = os.path.join(model_dir, "pattern_model")
        if os.path.exists(model_path):
            logger.info("Deploying pattern model to edge")
            eve_manager.deploy_edge_model("pattern_model", model_path)
    
    if services:
        # Print service information
        logger.info("\nServices started:")
        for service_name, thread, port in services:
            logger.info(f"  - {service_name}: http://localhost:{port}")
        
        # Keep main thread alive
        try:
            while True:
                time.sleep(1)
        except KeyboardInterrupt:
            logger.info("Shutting down...")
            sys.exit(0)
    else:
        logger.info("All tasks completed.")

if __name__ == "__main__":
    main()
