claude-dc-implementation/archive/requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
blis==1.2.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2020.6.20
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.4
confection==0.1.5
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
Flask==2.0.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.87.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
keras==2.14.0
kiwisolver==1.4.8
libclang==18.1.1
librosa==0.11.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
ml_dtypes==0.2.0
multidict==6.3.0
murmurhash==1.0.12
networkx==3.3
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.0
openai==1.70.0
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pooch==1.8.2
proto-plus==1.26.1
protobuf==4.25.3  # Compatible with tensorflow-rocm
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2022.1
PyWavelets==1.8.0
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
soundfile==0.13.1
spacy==3.8.4
spacy-legacy==3.0.12
srsly==2.5.1
starlette==0.46.1
streamlit==1.44.0
tensorboard==2.14.1
tensorflow-rocm==2.14.0.600
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
typer==0.15.2
typing_extensions==4.12.2
urllib3==1.26.5
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.0.1
wrapt==1.14.1
yarl==1.18.3


# Note: protobuf version pinned to 4.25.3 to maintain compatibility with tensorflow-rocm
# This prioritizes GPU acceleration for pattern processing over grpcio-status functionality
protobuf==4.25.3
# Brlapi, command-not-found, chrome-gnome-shell, cupshelpers, dbus-python, docker-py (docker), python-apt, system76driver, systemd-python, ubuntu-drivers-common, ufw, xkit system-level packages removed from requirements.txt
